Added new swift files

+++ Chart2D/test/sample/src/Graph2DVC.m
@@ -74,7 +74,7 @@
 }
 
 
-- (NSString *) graph2DView:(Graph2DView *) graph2DView yLabelAt:(NSInteger)y
+- (NSString *) graph2DView:(Graph2DView *) graph2DView yLabelAt:(double)y
 {
     return [NSString stringWithFormat:@"%ld", (long)y];
 }
